+++ services/blockchain-coordinator/jobs/delegates.js
@@ -21,8 +21,7 @@ module.exports = [
 	{
 		name: 'reload.delegates',
 		description: 'Keep the delegate list up-to-date',
-		// schedule: '*/5 * * * *', // Every 5 min
-		interval: 10,
+		schedule: '*/5 * * * *', // Every 5 min
 		init: async () => {
 			logger.debug('Initializing delegate cache...');
 			try {
